Refactoring complete for RucioClient
