+++ src/client/dids.ts
@@ -1,17 +1,21 @@
 import { streamData } from '../utils/restApiWrapper'
+import { DIDModel } from '../utils/models'
 
-export function search(scope: string, name: string, type: string) {
-    const url = '/dids/' + scope + '/dids/search'
-    return streamData(url, {
-        name: name,
-        type: type,
-    })
-}
-
-export function get_did(scope: string, name: string, dynamic: boolean) {
-    let url = '/dids/' + scope + '/' + name
-    if (dynamic === true) {
-        url += '?dynamic=' + dynamic
+class DID extends DIDModel {
+    public static search(scope: string, name: string, type: string) {
+        const url = '/dids/' + scope + '/dids/search'
+        return streamData(url, {
+            name: name,
+            type: type,
+        })
     }
-    return streamData(url)
+    public static get_did(scope: string, name: string, dynamic: boolean) {
+        let url = '/dids/' + scope + '/' + name
+        if (dynamic === true) {
+            url += '?dynamic=' + dynamic
+        }
+        return streamData(url)
+    }
 }
+
+export { DID }
