+++ src/client/index.ts
@@ -1,7 +1,9 @@
 import { Account } from './accounts'
 import { Rules } from './rules'
+import { DID } from './dids'
 
 export const RucioClient = {
     Account,
     Rules,
+    DID,
 }
