+++ src/utils/models.ts
@@ -60,3 +60,5 @@ export class RSEModel implements RSEType {
         this.total_quota = total_quota ?? ''
     }
 }
+
+export class AccountLimitModel implements AccountLimitType {}
